M2-2717 Copy font-face mixin from bourbon into Grana fonts file
This is a workaround to fix a problem with font loading caused by
the interaction between SASS sourcemaps and the URL resolver in
Webpack. Webpack does notlike that the "source" of the font file
URL is the mixin file rather than the actual source file. This
prevents the url resolver from being able to find the font file
as it is located somewhere completely different in the file tree.